REFACTOR: use GenericDef
