+++ base/src/main/java/org/aya/resolve/ModuleCallback.java
@@ -4,11 +4,12 @@ package org.aya.resolve;
 
 import kala.collection.immutable.ImmutableSeq;
 import org.aya.core.def.Def;
+import org.aya.core.def.GenericDef;
 import org.jetbrains.annotations.NotNull;
 
 /** CheckedTriConsumer */
 @FunctionalInterface
 public interface ModuleCallback<E extends Exception> {
-  void onModuleTycked(@NotNull ResolveInfo moduleResolve, @NotNull ImmutableSeq<Def> defs)
+  void onModuleTycked(@NotNull ResolveInfo moduleResolve, @NotNull ImmutableSeq<GenericDef> defs)
     throws E;
 }
